+++ src/oci/data_flow/models/application.py
@@ -118,6 +118,10 @@ class Application(object):
             The value to assign to the logs_bucket_uri property of this Application.
         :type logs_bucket_uri: str
 
+        :param metastore_id:
+            The value to assign to the metastore_id property of this Application.
+        :type metastore_id: str
+
         :param num_executors:
             The value to assign to the num_executors property of this Application.
         :type num_executors: int
@@ -173,6 +177,7 @@ class Application(object):
             'language': 'str',
             'lifecycle_state': 'str',
             'logs_bucket_uri': 'str',
+            'metastore_id': 'str',
             'num_executors': 'int',
             'owner_principal_id': 'str',
             'owner_user_name': 'str',
@@ -202,6 +207,7 @@ class Application(object):
             'language': 'language',
             'lifecycle_state': 'lifecycleState',
             'logs_bucket_uri': 'logsBucketUri',
+            'metastore_id': 'metastoreId',
             'num_executors': 'numExecutors',
             'owner_principal_id': 'ownerPrincipalId',
             'owner_user_name': 'ownerUserName',
@@ -230,6 +236,7 @@ class Application(object):
         self._language = None
         self._lifecycle_state = None
         self._logs_bucket_uri = None
+        self._metastore_id = None
         self._num_executors = None
         self._owner_principal_id = None
         self._owner_user_name = None
@@ -709,6 +716,30 @@ class Application(object):
         :type: str
         """
         self._logs_bucket_uri = logs_bucket_uri
+
+    @property
+    def metastore_id(self):
+        """
+        Gets the metastore_id of this Application.
+        The OCID of OCI Hive Metastore.
+
+
+        :return: The metastore_id of this Application.
+        :rtype: str
+        """
+        return self._metastore_id
+
+    @metastore_id.setter
+    def metastore_id(self, metastore_id):
+        """
+        Sets the metastore_id of this Application.
+        The OCID of OCI Hive Metastore.
+
+
+        :param metastore_id: The metastore_id of this Application.
+        :type: str
+        """
+        self._metastore_id = metastore_id
 
     @property
     def num_executors(self):