+++ src/oci/database/models/create_autonomous_database_from_backup_details.py
@@ -150,6 +150,11 @@ class CreateAutonomousDatabaseFromBackupDetails(CreateAutonomousDatabaseBase):
             The value to assign to the customer_contacts property of this CreateAutonomousDatabaseFromBackupDetails.
         :type customer_contacts: list[oci.database.models.CustomerContact]
 
+        :param autonomous_maintenance_schedule_type:
+            The value to assign to the autonomous_maintenance_schedule_type property of this CreateAutonomousDatabaseFromBackupDetails.
+            Allowed values for this property are: "EARLY", "REGULAR"
+        :type autonomous_maintenance_schedule_type: str
+
         :param autonomous_database_backup_id:
             The value to assign to the autonomous_database_backup_id property of this CreateAutonomousDatabaseFromBackupDetails.
         :type autonomous_database_backup_id: str
@@ -191,6 +196,7 @@ class CreateAutonomousDatabaseFromBackupDetails(CreateAutonomousDatabaseBase):
             'db_version': 'str',
             'source': 'str',
             'customer_contacts': 'list[CustomerContact]',
+            'autonomous_maintenance_schedule_type': 'str',
             'autonomous_database_backup_id': 'str',
             'clone_type': 'str'
         }
@@ -226,6 +232,7 @@ class CreateAutonomousDatabaseFromBackupDetails(CreateAutonomousDatabaseBase):
             'db_version': 'dbVersion',
             'source': 'source',
             'customer_contacts': 'customerContacts',
+            'autonomous_maintenance_schedule_type': 'autonomousMaintenanceScheduleType',
             'autonomous_database_backup_id': 'autonomousDatabaseBackupId',
             'clone_type': 'cloneType'
         }
@@ -260,6 +267,7 @@ class CreateAutonomousDatabaseFromBackupDetails(CreateAutonomousDatabaseBase):
         self._db_version = None
         self._source = None
         self._customer_contacts = None
+        self._autonomous_maintenance_schedule_type = None
         self._autonomous_database_backup_id = None
         self._clone_type = None
         self._source = 'BACKUP_FROM_ID'