+++ src/oci/database/models/create_autonomous_database_from_backup_timestamp_details.py
@@ -150,6 +150,11 @@ class CreateAutonomousDatabaseFromBackupTimestampDetails(CreateAutonomousDatabas
             The value to assign to the customer_contacts property of this CreateAutonomousDatabaseFromBackupTimestampDetails.
         :type customer_contacts: list[oci.database.models.CustomerContact]
 
+        :param autonomous_maintenance_schedule_type:
+            The value to assign to the autonomous_maintenance_schedule_type property of this CreateAutonomousDatabaseFromBackupTimestampDetails.
+            Allowed values for this property are: "EARLY", "REGULAR"
+        :type autonomous_maintenance_schedule_type: str
+
         :param autonomous_database_id:
             The value to assign to the autonomous_database_id property of this CreateAutonomousDatabaseFromBackupTimestampDetails.
         :type autonomous_database_id: str
@@ -195,6 +200,7 @@ class CreateAutonomousDatabaseFromBackupTimestampDetails(CreateAutonomousDatabas
             'db_version': 'str',
             'source': 'str',
             'customer_contacts': 'list[CustomerContact]',
+            'autonomous_maintenance_schedule_type': 'str',
             'autonomous_database_id': 'str',
             'timestamp': 'datetime',
             'clone_type': 'str'
@@ -231,6 +237,7 @@ class CreateAutonomousDatabaseFromBackupTimestampDetails(CreateAutonomousDatabas
             'db_version': 'dbVersion',
             'source': 'source',
             'customer_contacts': 'customerContacts',
+            'autonomous_maintenance_schedule_type': 'autonomousMaintenanceScheduleType',
             'autonomous_database_id': 'autonomousDatabaseId',
             'timestamp': 'timestamp',
             'clone_type': 'cloneType'
@@ -266,6 +273,7 @@ class CreateAutonomousDatabaseFromBackupTimestampDetails(CreateAutonomousDatabas
         self._db_version = None
         self._source = None
         self._customer_contacts = None
+        self._autonomous_maintenance_schedule_type = None
         self._autonomous_database_id = None
         self._timestamp = None
         self._clone_type = None